Moves test events helper
