+++ src/Popup/Popup.test.js
@@ -1,5 +1,5 @@
 import { Popup } from 'root';
-import { events } from 'root/utils/events';
+import { events } from 'root/src/lib/events';
 
 const {
   click,
